TESTS: add vm coverage tests
