+++ code/tests/testqvm2/test_ops.asm
