+++ code/tests/testqvm2/test_syscalls.asm
@@ -1,0 +1,3 @@
+code
+
+equ	test_report	-1
