+++ code/tests/testqvm2/test_syscalls.h
@@ -1,0 +1,6 @@
+#ifndef TEST_SYSCALLS_H
+#define TEST_SYSCALLS_H
+
+void test_report(int value);
+
+#endif
